email invite and displaying user works. need to work on remove users. organized some mor
